refactor(core): split config manager actor in different files
